moves samples to examples/ dir
This has mutliple benefits:

* The dependencies required for the examples no longer need to be
compiled when one is just using `afxdp-rs` as a library
* You can run the examples with `cargo run --release --example <example_name>`

I also made some slight changes to the `structopt` definition so that
`--link-name` is required (i.e. you longer need to check for it), and
`--copy`/`--zero-copy` now conflict with each other so you cannot use
them at the same time.