Clean-up XDP programs in examples on exit
Libbpf adds XDP programs when a XSK socket is created. Previously we
weren't cleaning those up properly on exit. This also exposed a problem
on start where the calls to libbpf to create a socket in each thread
raced and caused errors adding the BPF program. Fix this by moving the
socket creation to the main thread not the worker threads.